Fix major transformation issues
In the DirectX 9 docs, I'm sure they mention somewhere that they assume
vectors to be row-major and not column-major, but that's not how I built
this.

Transposing a couple of the transformation matrices, in particular the
LookAt and Projection matrices, which were the wrong way, so that
everything is column-major resolved the decade-old bug in the
transformations. Huzzah!

Also, the reason the spinning translucent triangle was orbiting a point
instead of spinning on its axis was actually an example of it working
slightly more correctly than not: the MODEL coordinates of the triangle
were slightly offset (y = 1 instead of 0) from the y-plane, and the
rotation was based on the origin / pivot point of the model's
coordinates.