Fix bug with directional lights
I think the docs I used are wrong for directional lights. Multiplying
but the view * world transforms when applying lighting means the light
seems to follow the surface from whatever its original orientation was,
i.e. the dot product of the surface normal and the light direction
doesn't change. That means that, e.g., if the light is from the left and
a surface starts rotating, it remains lit as if the light were orbiting
it.

Applying only the view transform matrix, like we do with point lights,
disconnects the light from each surface, meaning the dot product of the
lights direction and the surface normal, the lambert value, does change
and the surface can go in and out of shadow.